# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TournamentStatus = typeof TournamentStatus[keyof typeof TournamentStatus];


export const TournamentStatus = {
  registration: 'registration',
  active: 'active',
  completed: 'completed',
} as const;

export type TournamentFormat = typeof TournamentFormat[keyof typeof TournamentFormat];


export const TournamentFormat = {
  round_robin: 'round_robin',
  swiss: 'swiss',
} as const;

export type ParticipantStatus = typeof ParticipantStatus[keyof typeof ParticipantStatus];


export const ParticipantStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  pending: 'pending',
  ready: 'ready',
  completed: 'completed',
} as const;

export interface TournamentSummary {
  id: number;
  slug: string;
  name: string;
  status: TournamentStatus;
  participantCount: number;
  createdAt: string;
}

export interface TournamentInput {
  /**
     * @minLength 3
     * @maxLength 80
     */
  name: string;
  /** @nullable */
  registrationDeadline?: string | null;
  /**
     * @minimum 3
     * @maximum 128
     * @nullable
     */
  maxParticipants?: number | null;
}

export interface RegistrationInput {
  /**
     * @minLength 2
     * @maxLength 32
     */
  nickname: string;
}

export interface ParticipantUpdate {
  /**
     * @minLength 2
     * @maxLength 32
     */
  nickname?: string;
  status?: ParticipantStatus;
}

export interface MatchResultInput {
  /** @minimum 0 */
  player1Score: number;
  /** @minimum 0 */
  player2Score: number;
}

export interface Participant {
  id: number;
  nickname: string;
  status: ParticipantStatus;
  /** @nullable */
  seed: number | null;
  createdAt: string;
}

export interface TournamentMatch {
  id: number;
  stage: string;
  round: number;
  status: MatchStatus;
  targetScore: number;
  /** @nullable */
  player1Id: number | null;
  /** @nullable */
  player2Id: number | null;
  /** @nullable */
  player1Name: string | null;
  /** @nullable */
  player2Name: string | null;
  /** @nullable */
  player1Score: number | null;
  /** @nullable */
  player2Score: number | null;
}

export interface Standing {
  participantId: number;
  nickname: string;
  played: number;
  wins: number;
  losses: number;
  pointsFor: number;
  pointsAgainst: number;
  differential: number;
  rank: number;
}

/**
 * @nullable
 */
export type TournamentDetailFormat = typeof TournamentDetailFormat[keyof typeof TournamentDetailFormat] | null;


export const TournamentDetailFormat = {
  round_robin: 'round_robin',
  swiss: 'swiss',
} as const;

export interface TournamentDetail {
  id: number;
  slug: string;
  name: string;
  status: TournamentStatus;
  /** @nullable */
  format: TournamentDetailFormat;
  /** @nullable */
  swissRounds: number | null;
  /** @nullable */
  registrationDeadline?: string | null;
  /** @nullable */
  maxParticipants?: number | null;
  isOrganizer: boolean;
  participants: Participant[];
  matches: TournamentMatch[];
  standings: Standing[];
  createdAt: string;
}