# lib/api-zod/src/generated/types/matchResultInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchResultInput {
  /** @minimum 0 */
  player1Score: number;
  /** @minimum 0 */
  player2Score: number;
}