# lib/api-zod/src/generated/types/matchStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  pending: 'pending',
  ready: 'ready',
  completed: 'completed',
} as const;