# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantStatus } from './participantStatus';

export interface Participant {
  id: number;
  nickname: string;
  status: ParticipantStatus;
  /** @nullable */
  seed: number | null;
  createdAt: Date;
}