# lib/api-zod/src/generated/types/participantStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ParticipantStatus = typeof ParticipantStatus[keyof typeof ParticipantStatus];


export const ParticipantStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;