# lib/api-zod/src/generated/types/participantUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantStatus } from './participantStatus';

export interface ParticipantUpdate {
  /**
     * @minLength 2
     * @maxLength 32
     */
  nickname?: string;
  status?: ParticipantStatus;
}