# lib/api-zod/src/generated/types/tournamentDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Participant } from './participant';
import type { Standing } from './standing';
import type { TournamentDetailFormat } from './tournamentDetailFormat';
import type { TournamentMatch } from './tournamentMatch';
import type { TournamentStatus } from './tournamentStatus';

export interface TournamentDetail {
  id: number;
  slug: string;
  name: string;
  status: TournamentStatus;
  /** @nullable */
  format: TournamentDetailFormat;
  /** @nullable */
  swissRounds: number | null;
  /** @nullable */
  registrationDeadline?: Date | null;
  /** @nullable */
  maxParticipants?: number | null;
  isOrganizer: boolean;
  participants: Participant[];
  matches: TournamentMatch[];
  standings: Standing[];
  createdAt: Date;
}