# lib/api-zod/src/generated/types/tournamentFormat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TournamentFormat = typeof TournamentFormat[keyof typeof TournamentFormat];


export const TournamentFormat = {
  round_robin: 'round_robin',
  swiss: 'swiss',
} as const;