# lib/api-zod/src/generated/types/tournamentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TournamentInput {
  /**
     * @minLength 3
     * @maxLength 80
     */
  name: string;
  /** @nullable */
  registrationDeadline?: Date | null;
  /**
     * @minimum 3
     * @maximum 128
     * @nullable
     */
  maxParticipants?: number | null;
}