# lib/api-zod/src/generated/types/tournamentMatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from './matchStatus';

export interface TournamentMatch {
  id: number;
  stage: string;
  round: number;
  status: MatchStatus;
  targetScore: number;
  /** @nullable */
  player1Id: number | null;
  /** @nullable */
  player2Id: number | null;
  /** @nullable */
  player1Name: string | null;
  /** @nullable */
  player2Name: string | null;
  /** @nullable */
  player1Score: number | null;
  /** @nullable */
  player2Score: number | null;
}