# lib/api-zod/src/generated/types/tournamentStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TournamentStatus = typeof TournamentStatus[keyof typeof TournamentStatus];


export const TournamentStatus = {
  registration: 'registration',
  active: 'active',
  completed: 'completed',
} as const;