# lib/api-zod/src/generated/types/tournamentSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentStatus } from './tournamentStatus';

export interface TournamentSummary {
  id: number;
  slug: string;
  name: string;
  status: TournamentStatus;
  participantCount: number;
  createdAt: Date;
}